+++ K-Content/src/main/resources/templates/cms/cntnt/contentDetail.html
@@ -1,89 +1,89 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>CSS Grid</title>
-    <link rel="stylesheet" href="/css/content/default.css">
-    <link rel="stylesheet" href="/css/content/ui-ie.css">
-    <link rel="stylesheet" href="/css/content/cntntDetailGrid.css">
-    <link rel="stylesheet"
-          href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@24,400,0,0"/>
-</head>
-<body>
+<!--<!DOCTYPE html>-->
+<!--<html>-->
+<!--<head>-->
+<!--    <meta charset="UTF-8">-->
+<!--    <meta name="viewport" content="width=device-width, initial-scale=1.0">-->
+<!--    <meta http-equiv="X-UA-Compatible" content="ie=edge">-->
+<!--    <title>CSS Grid</title>-->
+<!--    <link rel="stylesheet" href="/css/content/default.css">-->
+<!--    <link rel="stylesheet" href="/css/content/ui-ie.css">-->
+<!--    <link rel="stylesheet" href="/css/content/cntntDetailGrid.css">-->
+<!--    <link rel="stylesheet"-->
+<!--          href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@24,400,0,0"/>-->
+<!--</head>-->
+<!--<body>-->
 
-<div class="page">
-    <header class="header">
-        <h1 class="website-title">k-spectrum</h1>
-    </header>
-    <!--    main-->
-    <div class="main-content">
-        <input id="cntntId" type="hidden" th:value="${content.cntntId}">
-        <p>
-        <div id="iframe" th:src="${content.cntntUrl}" alt="">
-        </div>
-        </p>
-        <div class="item category-caption iframe-title" id="title">
-            <h3 style="text-align: center; " th:text="${content.cntntTitle}">title</h3>
-        </div>
-        <button th:text="${category.getCommonCodeVal()}"></button>
+<!--<div class="page">-->
+<!--    <header class="header">-->
+<!--        <h1 class="website-title">k-spectrum</h1>-->
+<!--    </header>-->
+<!--    &lt;!&ndash;    main&ndash;&gt;-->
+<!--    <div class="main-content">-->
+<!--        <input id="cntntId" type="hidden" th:value="${content.cntntId}">-->
+<!--        <p>-->
+<!--        <div id="iframe" th:src="${content.cntntUrl}" alt="">-->
+<!--        </div>-->
+<!--        </p>-->
+<!--        <div class="item category-caption iframe-title" id="title">-->
+<!--            <h3 style="text-align: center; " th:text="${content.cntntTitle}">title</h3>-->
+<!--        </div>-->
+<!--        <button th:text="${category.getCommonCodeVal()}"></button>-->
 
-        <div class="item keyword">
-            <div class="keywordItem" th:each="keyword : ${keywordList}">
-                <h2 th:text="${keyword}"></h2>
-            </div>
-        </div>
+<!--        <div class="item keyword">-->
+<!--            <div class="keywordItem" th:each="keyword : ${keywordList}">-->
+<!--                <h2 th:text="${keyword}"></h2>-->
+<!--            </div>-->
+<!--        </div>-->
 
-        <div class="item goods-box">
-            <div th:each="goods : ${goodsJFileList}" id="goods">
-                <a th:href="${goods.goodsPurchsLink}">
-                    <div class="card" style="width: 18rem; height: 240px">
-                        <div style="width:18rem; height:140px">
-                            <img style="width:100%; height:100%"
-                                 th:src="@{/img/goods/{image}(image=${goods.goodsFileId})}" class="card-img-top"
-                                 alt="...">
-                        </div>
-                        <div style="width:18rem; height:100px;border:1px solid">
-                            <h5 class="text-center" th:text="${goods.goodsName}">Card title</h5>
-                            <p class="text-center" th:text="${goods.goodsPrice} + ' 원'"></p>
-                            <p class="text-center" th:text="${goods.goodsBrand}"></p>
-                        </div>
-                    </div>
-                </a>
-            </div>
+<!--        <div class="item goods-box">-->
+<!--            <div th:each="goods : ${goodsJFileList}" id="goods">-->
+<!--                <a th:href="${goods.goodsPurchsLink}">-->
+<!--                    <div class="card" style="width: 18rem; height: 240px">-->
+<!--                        <div style="width:18rem; height:140px">-->
+<!--                            <img style="width:100%; height:100%"-->
+<!--                                 th:src="@{/img/goods/{image}(image=${goods.goodsFileId})}" class="card-img-top"-->
+<!--                                 alt="...">-->
+<!--                        </div>-->
+<!--                        <div style="width:18rem; height:100px;border:1px solid">-->
+<!--                            <h5 class="text-center" th:text="${goods.goodsName}">Card title</h5>-->
+<!--                            <p class="text-center" th:text="${goods.goodsPrice} + ' 원'"></p>-->
+<!--                            <p class="text-center" th:text="${goods.goodsBrand}"></p>-->
+<!--                        </div>-->
+<!--                    </div>-->
+<!--                </a>-->
+<!--            </div>-->
 
-        </div>
-    </div>
-    <!--    mainend-->
-    <hr>
+<!--        </div>-->
+<!--    </div>-->
+<!--    &lt;!&ndash;    mainend&ndash;&gt;-->
+<!--    <hr>-->
 
-    <section class="recommand">
-        <h2>추천 콘텐츠</h2>
-        <ul class="card-list" id="card-list">
-        </ul>
-    </section>
+<!--    <section class="recommand">-->
+<!--        <h2>추천 콘텐츠</h2>-->
+<!--        <ul class="card-list" id="card-list">-->
+<!--        </ul>-->
+<!--    </section>-->
 
-    <aside class="secondary secondary-a" th:insert="~{/include/admin-sideBar :: adminSideBar}">
-    </aside>
+<!--    <aside class="secondary secondary-a" th:insert="~{/include/admin-sideBar :: adminSideBar}">-->
+<!--    </aside>-->
 
-    <div class="admin-button">
-        <button onclick="updateCntnt()">수정</button>
-        <button onclick="deleteCntnt()">삭제</button>
-    </div>
+<!--    <div class="admin-button">-->
+<!--        <button onclick="updateCntnt()">수정</button>-->
+<!--        <button onclick="deleteCntnt()">삭제</button>-->
+<!--    </div>-->
 
-    <footer class="footer" th:insert="~{/include/admin-footer :: admin-footer}">
-    </footer>
+<!--    <footer class="footer" th:insert="~{/include/admin-footer :: admin-footer}">-->
+<!--    </footer>-->
 
-</div>
+<!--</div>-->
 
 
-<input class="item" id="trendQueryList" type="hidden" th:value="${trendQueryList}"
-       style="display: none">
-<input type="hidden" id="contentId" th:name="${targetContentIdF}" th:value="${content.getCntntUrl()}">
-<script src="/vendor/jquery/jquery.min.js"></script>
-<script src="/js/cntnt/cntntDetail.js" defer></script>
+<!--<input class="item" id="trendQueryList" type="hidden" th:value="${trendQueryList}"-->
+<!--       style="display: none">-->
+<!--<input type="hidden" id="contentId" th:name="${targetContentIdF}" th:value="${content.getCntntUrl()}">-->
+<!--<script src="/vendor/jquery/jquery.min.js"></script>-->
+<!--<script src="/js/cntnt/cntntDetail.js" defer></script>-->
 
-</body>
+<!--</body>-->
 
-</html>
+<!--</html>-->
