+++ K-Content/src/main/resources/templates/cms/cntnt/newcontentRecom.html
@@ -1,47 +1,47 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>CSS Grid</title>
-    <link rel="stylesheet" href="/css/content/default.css">
-    <link rel="stylesheet" href="/css/content/ui-ie.css">
-    <link rel="stylesheet" href="/css/content/cntntManage.css">
-    <link rel="stylesheet"
-          href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@24,400,0,0"/>
-</head>
-<body>
-<input type="checkbox" id="modal-switch">
+<!--<!DOCTYPE html>-->
+<!--<html>-->
+<!--<head>-->
+<!--    <meta charset="UTF-8">-->
+<!--    <meta name="viewport" content="width=device-width, initial-scale=1.0">-->
+<!--    <meta http-equiv="X-UA-Compatible" content="ie=edge">-->
+<!--    <title>CSS Grid</title>-->
+<!--    <link rel="stylesheet" href="/css/content/default.css">-->
+<!--    <link rel="stylesheet" href="/css/content/ui-ie.css">-->
+<!--    <link rel="stylesheet" href="/css/content/cntntManage.css">-->
+<!--    <link rel="stylesheet"-->
+<!--          href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@24,400,0,0"/>-->
+<!--</head>-->
+<!--<body>-->
+<!--<input type="checkbox" id="modal-switch">-->
 
-<div class="page">
-    <header class="header">
-        <h1 class="website-title">k-spectrum</h1>
-        <form class="search-form" id="search-form" action="/cs/recomm" method="get">
-            <input type="search" id="search-input" th:name="search">
-            <input id="search-button" onclick="searchButton()" type="button"  value="찾기">
-        </form>
-    </header>
+<!--<div class="page">-->
+<!--    <header class="header">-->
+<!--        <h1 class="website-title">k-spectrum</h1>-->
+<!--        <form class="search-form" id="search-form" action="/cs/recomm" method="get">-->
+<!--            <input type="search" id="search-input" th:name="search">-->
+<!--            <input id="search-button" onclick="searchButton()" type="button"  value="찾기">-->
+<!--        </form>-->
+<!--    </header>-->
 
-    <section class="primary">
-        <ul class="card-list" id="card-list">
+<!--    <section class="primary">-->
+<!--        <ul class="card-list" id="card-list">-->
 
-        </ul>
-    </section>
+<!--        </ul>-->
+<!--    </section>-->
 
-    <aside class="secondary secondary-a" th:insert="~{/include/admin-sideBar :: adminSideBar}">
-    </aside>
-    <!-- </div> -->
+<!--    <aside class="secondary secondary-a" th:insert="~{/include/admin-sideBar :: adminSideBar}">-->
+<!--    </aside>-->
+<!--    &lt;!&ndash; </div> &ndash;&gt;-->
 
-    <footer class="footer" th:insert="~{/include/admin-footer :: admin-footer}">
-    </footer>
+<!--    <footer class="footer" th:insert="~{/include/admin-footer :: admin-footer}">-->
+<!--    </footer>-->
 
-</div>
+<!--</div>-->
 
 
-<script src="/vendor/jquery/jquery.min.js"></script>
-<script src="/js/cntnt/cntntRecom.js"></script>
+<!--<script src="/vendor/jquery/jquery.min.js"></script>-->
+<!--<script src="/js/cntnt/cntntRecom.js"></script>-->
 
-</body>
+<!--</body>-->
 
-</html>
+<!--</html>-->
